Move reconnect flow to TerminalManager; drop stale actions.py duplicates
The controlled-reconnect flow (prompt_reconnect, response handling,
_reconnect_terminal, _show_reconnect_error) now lives on TerminalManager;
window.py keeps only the _is_controlled_reconnect flag terminal.py reads
via the root window. The prompt and error popups use Adw.AlertDialog /
the window's _error_dialog helper instead of raw Gtk.MessageDialog.

Also deletes the last stale shadowed copies in actions.py
(on_edit_connection_action, on_delete_connection_action,
on_open_in_system_terminal_action) — MainWindow's own definitions had
superseded them.
