+++ sshpilot/actions.py
@@ -283,76 +283,6 @@ class WindowActions:
             except Exception:
                 pass
 
-    def on_edit_connection_action(self, action, param=None):
-        """Handle edit connection action from context menu"""
-        try:
-            connection = getattr(self, '_context_menu_connection', None)
-            if connection is None:
-                # Fallback to selected row if any
-                row = self.connection_list.get_selected_row()
-                connection = getattr(row, 'connection', None) if row else None
-            if connection is None:
-                return
-            self.show_connection_dialog(connection)
-        except Exception as e:
-            logger.error(f"Failed to edit connection: {e}")
-
-    def on_delete_connection_action(self, action, param=None):
-        """Handle delete connection action from context menu"""
-        try:
-            connection = getattr(self, '_context_menu_connection', None)
-            if connection is None:
-                # Fallback to selected row if any
-                row = self.connection_list.get_selected_row()
-                connection = getattr(row, 'connection', None) if row else None
-            if connection is None:
-                return
-
-            # Use the same logic as the button click handler
-            # If host has active connections/tabs, warn about closing them first
-            has_active_terms = bool(self.connection_to_terminals.get(connection, []))
-            if getattr(connection, 'is_connected', False) or has_active_terms:
-                dialog = Adw.MessageDialog(
-                    transient_for=self,
-                    modal=True,
-                    heading=_('Remove host?'),
-                    body=_('Close connections and remove host?')
-                )
-                dialog.add_response('cancel', _('Cancel'))
-                dialog.add_response('close_remove', _('Close and Remove'))
-                dialog.set_response_appearance('close_remove', Adw.ResponseAppearance.DESTRUCTIVE)
-                dialog.set_default_response('close')
-                dialog.set_close_response('cancel')
-            else:
-                # Simple delete confirmation when not connected
-                dialog = Adw.MessageDialog.new(self, _('Delete Connection?'),
-                                             _('Are you sure you want to delete "{}"?').format(connection.nickname))
-                dialog.add_response('cancel', _('Cancel'))
-                dialog.add_response('delete', _('Delete'))
-                dialog.set_response_appearance('delete', Adw.ResponseAppearance.DESTRUCTIVE)
-                dialog.set_default_response('cancel')
-                dialog.set_close_response('cancel')
-
-            dialog.connect('response', self.on_delete_connection_response, connection)
-            dialog.present()
-        except Exception as e:
-            logger.error(f"Failed to delete connection: {e}")
-
-    def on_open_in_system_terminal_action(self, action, param=None):
-        """Handle open in system terminal action from context menu"""
-        try:
-            connection = getattr(self, '_context_menu_connection', None)
-            if connection is None:
-                # Fallback to selected row if any
-                row = self.connection_list.get_selected_row()
-                connection = getattr(row, 'connection', None) if row else None
-            if connection is None:
-                return
-
-            self.open_in_system_terminal(connection)
-        except Exception as e:
-            logger.error(f"Failed to open in system terminal: {e}")
-
     def on_wake_on_lan_action(self, action, param=None):
         """Send Wake-on-LAN magic packets for the targeted connections.
