+++ sshpilot/window.py
@@ -5,7 +5,6 @@ Primary UI with connection list, tabs, and terminal management
 
 from __future__ import annotations
 
-import asyncio
 import copy
 import os
 import logging
@@ -4894,7 +4893,7 @@ class MainWindow(Adw.ApplicationWindow, WindowBroadcastMixin, WindowSessionMixin
                 if is_connected and terminal is not None:
                     # Store the terminal in the connection for later use
                     old_connection._terminal_instance = terminal
-                    self._prompt_reconnect(old_connection)
+                    self.terminal_manager.prompt_reconnect(old_connection)
                 _complete_save(True)
 
             else:
@@ -4919,169 +4918,3 @@ class MainWindow(Adw.ApplicationWindow, WindowBroadcastMixin, WindowSessionMixin
             self.rebuild_connection_list()
         except Exception:
             pass
-    def _prompt_reconnect(self, connection):
-        """Show a dialog asking if user wants to reconnect with new settings"""
-        dialog = Gtk.MessageDialog(
-            transient_for=self,
-            modal=True,
-            message_type=Gtk.MessageType.QUESTION,
-            buttons=Gtk.ButtonsType.YES_NO,
-            text=_("Settings Changed"),
-            secondary_text=_("The connection settings have been updated.\n"
-                           "Would you like to reconnect with the new settings?")
-        )
-        dialog.connect("response", self._on_reconnect_response, connection)
-        dialog.present()
-    
-    def _on_reconnect_response(self, dialog, response_id, connection):
-        """Handle response from reconnect prompt"""
-        dialog.destroy()
-        
-        # Only proceed if user clicked Yes and the connection is still active
-        if response_id != Gtk.ResponseType.YES or connection not in self.active_terminals:
-            # Clean up the stored terminal instance if it exists
-            if hasattr(connection, '_terminal_instance'):
-                delattr(connection, '_terminal_instance')
-            return
-            
-        # Get the terminal instance either from active_terminals or the stored instance
-        terminal = self.active_terminals.get(connection) or getattr(connection, '_terminal_instance', None)
-        if not terminal:
-            logger.warning("No terminal instance found for reconnection")
-            return
-
-        # Ensure the tab for this connection is focused so the user can
-        # observe the reconnection process even if another tab was
-        # previously active.
-        try:
-            self._focus_most_recent_tab(connection)
-        except Exception:
-            pass
-
-        # Set controlled reconnect flag
-        self._is_controlled_reconnect = True
-
-
-        
-        try:
-            # Disconnect first (defer to avoid blocking)
-            logger.debug("Disconnecting terminal before reconnection")
-            def _safe_disconnect():
-                try:
-                    terminal.disconnect()
-                    logger.debug("Terminal disconnected, scheduling reconnect")
-                    # Store the connection temporarily in active_terminals if not present
-                    if connection not in self.active_terminals:
-                        self.active_terminals[connection] = terminal
-                    # Reconnect after disconnect completes
-                    GLib.timeout_add(1000, self._reconnect_terminal, connection)  # Increased delay
-                except Exception as e:
-                    logger.error(f"Error during disconnect: {e}")
-                    GLib.idle_add(self._show_reconnect_error, connection, str(e))
-                return False
-            
-            # Defer disconnect to avoid blocking the UI thread
-            GLib.idle_add(_safe_disconnect)
-            
-        except Exception as e:
-            logger.error(f"Error during reconnection: {e}")
-            # Remove from active terminals if reconnection fails
-            if connection in self.active_terminals:
-                del self.active_terminals[connection]
-                
-            # Show error to user
-            error_dialog = Gtk.MessageDialog(
-                transient_for=self,
-                modal=True,
-                message_type=Gtk.MessageType.ERROR,
-                buttons=Gtk.ButtonsType.OK,
-                text=_("Reconnection Failed"),
-                secondary_text=_("Failed to reconnect with the new settings. Please try connecting again manually.")
-            )
-            error_dialog.present()
-            
-        finally:
-            # Clean up the stored terminal instance
-            if hasattr(connection, '_terminal_instance'):
-                delattr(connection, '_terminal_instance')
-                
-            # Reset the flag after a delay to ensure it's not set during normal operations
-            GLib.timeout_add(1000, self._reset_controlled_reconnect)
-    
-    def _reset_controlled_reconnect(self):
-        """Reset the controlled reconnect flag"""
-        self._is_controlled_reconnect = False
-    
-    def _reconnect_terminal(self, connection):
-        """Reconnect a terminal with updated connection settings"""
-        if connection not in self.active_terminals:
-            logger.warning(f"Connection {connection.nickname} not found in active terminals")
-            return False  # Don't repeat the timeout
-
-        terminal = self.active_terminals[connection]
-
-        try:
-            logger.debug(f"Attempting to reconnect terminal for {connection.nickname}")
-
-            # Rebuild the SSH command using the latest configuration so that
-            # options resolved via ssh -G are honored for the reconnect.
-            # Plugin protocols rebuild statelessly in build_spawn() instead.
-            if getattr(connection, 'protocol', 'ssh') == 'ssh':
-                try:
-                    loop = asyncio.get_event_loop()
-                    # Native-only (connect() delegates to native_connect()).
-                    if hasattr(connection, 'native_connect'):
-                        connect_coro = connection.native_connect()
-                    else:
-                        connect_coro = connection.connect()
-                    if loop.is_running():
-                        future = asyncio.run_coroutine_threadsafe(connect_coro, loop)
-                        future.result()
-                    else:
-                        loop.run_until_complete(connect_coro)
-                except Exception as prep_err:
-                    logger.error(
-                        "Failed to prepare SSH command before reconnect: %s",
-                        prep_err,
-                    )
-                    GLib.idle_add(self._show_reconnect_error, connection, str(prep_err))
-                    return False
-
-            # Reconnect with new settings
-            if not terminal._connect_ssh():
-                logger.error("Failed to reconnect with new settings")
-                # Show error to user
-                GLib.idle_add(self._show_reconnect_error, connection)
-                return False
-                
-            logger.info(f"Successfully reconnected terminal for {connection.nickname}")
-            
-        except Exception as e:
-            logger.error(f"Error reconnecting terminal: {e}", exc_info=True)
-            GLib.idle_add(self._show_reconnect_error, connection, str(e))
-            
-        return False  # Don't repeat the timeout
-        
-    def _show_reconnect_error(self, connection, error_message=None):
-        """Show an error message when reconnection fails"""
-        # Remove from active terminals if reconnection fails
-        if connection in self.active_terminals:
-            del self.active_terminals[connection]
-            
-        # Update UI to show disconnected state
-        for row in self._rows_for_connection(connection):
-            row.update_status()
-        
-        # Show error dialog
-        error_dialog = Gtk.MessageDialog(
-            transient_for=self,
-            modal=True,
-            message_type=Gtk.MessageType.ERROR,
-            buttons=Gtk.ButtonsType.OK,
-            text=_("Reconnection Failed"),
-            secondary_text=error_message or _("Failed to reconnect with the new settings. Please try connecting again manually.")
-        )
-        error_dialog.present()
-        
-        # Clean up the dialog when closed
-        error_dialog.connect("response", lambda d, r: d.destroy())
